+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+<!-- cSpell:ignore positon -->
+
 # Changelog
 
-## v1.0.3 2024-06
+- Add `eslint` formatting configuration
+- Add `cSpell` spelling check configuration
+- Add `typedoc` documentation generation configuration
+- Add documentation
+- Adjust `typedoc-plugin-markdown` configuration
 
 ## v1.0.2 2024-06-04
 
@@ -70,7 +76,7 @@
 ## 0.8.7 2023-10-24
 
 * [Deprecated `loaded-protyle` use `loaded-protyle-static` instead](https://github.com/siyuan-note/siyuan/issues/9468)
-* [Fix event defail for `loaded-protyle`](https://github.com/siyuan-note/petal/pull/18)
+* [Fix event defined for `loaded-protyle`](https://github.com/siyuan-note/petal/pull/18)
 * [Add plugin event bus `paste`](https://github.com/siyuan-note/siyuan/issues/9452)
 
 ## 0.8.6 2023-10-17